# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FXREBATE PROJECT",
        "description": "REACT Progressive Web App / Stand on the shoulders of giants",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/brokers/": {
            "get": {
                "tags": [
                    "Broker"
                ],
                "summary": "Get all brokers",
                "operationId": "b26d07fcdea5773e6cd7ddda550506de",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "These request do not match our records"
                    }
                }
            }
        },
        "/api/v1/brokers": {
            "post": {
                "tags": [
                    "Broker"
                ],
                "summary": "Add a new broker",
                "operationId": "60a61cac11e93f5f9cb5bc38d2d7948c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "registration_language": {
                                        "type": "string",
                                        "example": "en"
                                    },
                                    "registration_zone": {
                                        "type": "string",
                                        "example": "US"
                                    },
                                    "broker_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "These credentials do not match our records"
                    }
                }
            }
        },
        "/api/v1/broker/{id}": {
            "get": {
                "tags": [
                    "Broker"
                ],
                "summary": "Show a broker",
                "operationId": "fb382c8caef7d7e1be19438f21cd0f2b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Broker ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "These request do not match our records"
                    }
                }
            },
            "put": {
                "tags": [
                    "Broker"
                ],
                "summary": "Update broker",
                "operationId": "17715e762a5918b1fce4ab878b538f60",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Broker ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "registration_language": {
                                        "type": "string",
                                        "example": "en"
                                    },
                                    "registration_zone": {
                                        "type": "string",
                                        "example": "US"
                                    },
                                    "broker_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "These credentials do not match our records"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Broker"
                ],
                "summary": "Delete broker",
                "operationId": "084f3abaae194c1242baf66a2d03e26d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Broker ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
    },
...
}